Split `crate::point::{build,apply}` into separate files.
